fix(auth): 게시글 요청값 검증 추가
